+++ src/store/plugin/saga.ts
@@ -18,8 +18,8 @@ function* handleGetPluginDetails(action: IActionTypeParam) {
   try {
     const item: IPluginItem = action.payload;
 
-    const res = yield call(ChrisModel.fetchRequest, item.descendants); // Get descendants first:
     yield call(ChrisModel.fetchRequest, item.url);
+    const res = yield call(ChrisModel.fetchRequest, item.descendants); // Get descendants first:
 
     if (res.error) {
       console.error(res.error);
